+++ ios/ApteligentExample/Assets.xcassets/AppIcon.appiconset/Contents.json
